# common-platform-accounting-gl-sdk/target/classes/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-accounting-gl
  description: General Ledger Accounting Management
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.catalis.vc/general-ledger
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/gl-accounts/{id}:
    get:
      tags:
        - GL Accounts
      summary: Get GL Account by ID
      operationId: getById_3
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GlAccountDTO'
    put:
      tags:
        - GL Accounts
      summary: Update GL Account
      operationId: update_3
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GlAccountDTO'
        required: true
      responses:
        '200':
          description: GL Account updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GlAccountDTO'
    delete:
      tags:
        - GL Accounts
      summary: Delete GL Account
      operationId: delete_3
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: GL Account deleted
...
  /api/v1/gl-accounts:
    post:
      tags:
        - GL Accounts
      summary: Create GL Account
      operationId: create_3
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GlAccountDTO'
        required: true
      responses:
        '201':
          description: GL Account created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GlAccountDTO'
  /api/v1/gl-accounts/{parentId}/children:
    get:
      tags:
        - GL Accounts
      summary: List child GL Accounts for a given parent
      operationId: findChildren
      parameters:
        - name: parentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.code
          in: query
          description: Exact filter for code
          required: false
          schema:
            type: string
        - name: filters.name
          in: query
          description: Exact filter for name
          required: false
          schema:
            type: string
        - name: filters.type
          in: query
          description: Exact filter for type
          required: false
          schema:
            type: string
            enum:
              - ASSET
              - LIABILITY
              - INCOME
              - EXPENSE
              - EQUITY
        - name: filters.parentId
          in: query
          description: Exact filter for parent id
          required: false
          schema:
            type: string
        - name: filters.isLeaf
          in: query
          description: Exact filter for is leaf
          required: false
          schema:
            type: boolean
        - name: filters.active
          in: query
          description: Exact filter for active
          required: false
          schema:
            type: boolean
        - name: filters.currency
          in: query
          description: Exact filter for currency
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseGlAccountDTO'
    post:
      tags:
        - GL Accounts
      summary: Create a child GL Account under a parent
      operationId: createChild
      parameters:
        - name: parentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GlAccountDTO'
        required: true
      responses:
        '201':
          description: Child GL Account created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GlAccountDTO'
...
  /api/v1/gl-accounts/:
    get:
      tags:
        - GL Accounts
      summary: Search/Filter GL Accounts
      operationId: search_3
      parameters:
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.code
          in: query
          description: Exact filter for code
          required: false
          schema:
            type: string
        - name: filters.name
          in: query
          description: Exact filter for name
          required: false
          schema:
            type: string
        - name: filters.type
          in: query
          description: Exact filter for type
          required: false
          schema:
            type: string
            enum:
              - ASSET
              - LIABILITY
              - INCOME
              - EXPENSE
              - EQUITY
        - name: filters.parentId
          in: query
          description: Exact filter for parent id
          required: false
          schema:
            type: string
        - name: filters.isLeaf
          in: query
          description: Exact filter for is leaf
          required: false
          schema:
            type: boolean
        - name: filters.active
          in: query
          description: Exact filter for active
          required: false
          schema:
            type: boolean
        - name: filters.currency
          in: query
          description: Exact filter for currency
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseGlAccountDTO'
components:
  schemas:
...
    GlAccountDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        code:
          type: string
        name:
          type: string
        type:
          type: string
          enum:
            - ASSET
            - LIABILITY
            - INCOME
            - EXPENSE
            - EQUITY
        parentId:
          type: string
          format: uuid
        active:
          type: boolean
        currency:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        leaf:
          type: boolean
...
    PaginationResponseGlAccountDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/GlAccountDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.